Fix for Github issue #6
When using the USPS address validate api, it was possible to send in a
state name and it would be abbreviated to the first 2 letters in the
string. For example, Connecticut would be shortened to CO, which is the
abreviation for colorado.

Now when address objects are created, the constructor will try and match
a spelt out state name to the proper state abbreviation. If a state
abbreviation is inserted, then the wrapper will just take the input and
continue